int support

+++ core/src/commonMain/kotlin/org/kobjects/tantilla2/core/runtime/Range.kt
@@ -3,12 +3,12 @@ package org.kobjects.tantilla2.core.runtime
 import org.kobjects.tantilla2.core.Type
 import org.kobjects.tantilla2.core.Typed
 
-data class Range(val start: Double, val end: Double) : Typed, Iterable<Double> {
+data class Range(val start: Long, val end: Long) : Typed, Iterable<Long> {
     override val type: Type
         get() = RangeType
 
-    override fun iterator(): Iterator<Double> {
-        return object : Iterator<Double> {
+    override fun iterator(): Iterator<Long> {
+        return object : Iterator<Long> {
             var current = start
 
             override fun hasNext() = current < end
